requirements.txt
# Main package - includes all necessary dependencies:
# - geopandas, pandas, shapely, matplotlib
# - leafmap, overturemaps
# - rasterio, rioxarray
# - torch, torchgeo
# - and more...
geoai-py

# Geospatial geometry operations used directly by the backend
pandas>=2.0.0
shapely>=2.0.0
rasterio>=1.3.0

# Web server and API
flask==2.3.2
flask-cors==4.0.0

# Image processing
pillow==10.0.0
numpy==1.24.3

# HTTP client
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Direct PostgreSQL bulk import for Da Nang Overture buildings
psycopg[binary]>=3.2.0

# Elasticsearch + MiniLM property search infrastructure
elasticsearch>=8.15.0,<9.0.0
sentence-transformers>=3.0.0
